Add files via upload
In this task, I focused on evaluating the effectiveness of a targeted intervention in retail stores using a trial and control store experiment. This was necessary to test the effectiveness of an intervention, comparing trial and control stores, and using data analysis to measure the success of the changes implemented.